# rap-lp-openapi.yaml
...
openapi: 3.1.0
info:
  title: RAP-LP API
  version: 1.0.11
  summary: API for validating OpenAPI specifications against the Swedish REST API profile.
  description: |
    RAP-LP (REST API-profil Lint Processor) validates an OpenAPI specification against
    the national REST API profile and additional RAP-LP design rules.

    RAP-LP validates structure, semantics and design rules, and can also generate an
    Excel report from a precomputed validation report.

    This specification describes the linter API itself.

    Begransningar och problem: URL-baserad validering kan vara avstangd via serverkonfiguration,
    och generate-report returnerar en binar Excel-fil medan ovriga resurser returnerar JSON.
  contact:
    name: Digg
    url: https://github.com/diggsweden/rest-api-profil-lint-processor/
    email: apiteam@digg.se
  license:
    name: EUPL-1.2
    url: https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
  termsOfService: https://www.dataportal.se/terms
...
servers:
  - url: https://raplp.digg.se/api/v1/
    description: Production
  - url: https://raplp.sandbox.digg.se/api/v1/
    description: Sandbox
  - url: http://localhost:3000/api/v1/
    description: Local development
...
security: []
paths:
...
  /validation/url:
    post:
      tags:
        - validation
      summary: Validate a remotely hosted OpenAPI specification.
      description: |
        Fetches an OpenAPI specification from a remote URL and validates it.

        Redirects are not allowed. The requested URL is checked before fetching,
        then the downloaded OpenAPI document is parsed and strictly validated before
        RAP-LP rule validation is executed.

        This feature can be disabled by server configuration. When disabled, the API returns 409 Conflict.
      operationId: validateUrl
      requestBody:
        required: true
        description: Remote URL to the OpenAPI specification and optional validation settings.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UrlValidationRequest'
            examples:
              remoteSpecification:
                value:
                  url: https://api.example.se/foretagsinformation/v2/openapi.yaml
                  strict: true
                  categories:
                    - DokRules
                    - UfnRules
      responses:
        '200':
          description: Validation completed without rule violations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlValidationSuccessResponse'
              examples:
                success:
                  value:
                    ok: true
                    stage: rule-engine
                    format: yaml
                    payload:
                      result: []
                      report:
                        - Notering: Godkanda regler - RAP-LP
                          regler:
                            - id: DOK.03
                              område: Dokumentation
                              status: OK
        '400':
          description: Invalid request, invalid URL, invalid remote response, strict validation error or rule validation error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
              examples:
                missingUrl:
                  value:
                    type: https://raplp.digg.se/problems/bad-request
                    title: Invalid Request
                    status: 400
                    detail: 'Required field missing: url'
                    instance: /api/v1/validation/url
                urlPatternRejected:
                  value:
                    type: https://raplp.digg.se/problems/bad-request
                    title: Invalid Request
                    status: 400
                    detail: The requested address did not meet the allowed URL pattern. Please contact your administrator if you believe this is a mistake.
                    instance: /api/v1/validation/url
                invalidUrlFormat:
                  value:
                    type: https://raplp.digg.se/problems/bad-request
                    title: Invalid Request
                    status: 400
                    detail: Invalid URL format.
                    instance: /api/v1/validation/url
                invalidProtocol:
                  value:
                    type: https://raplp.digg.se/problems/bad-request
                    title: Invalid Request
                    status: 400
                    detail: Only HTTPS and HTTP protocols are allowed.
                    instance: /api/v1/validation/url
                urlFetchFailed:
                  value:
                    type: https://raplp.digg.se/problems/bad-request
                    title: Invalid Request
                    status: 400
                    detail: The requested URL could not be retrieved. Redirects are not allowed.
                    instance: /api/v1/validation/url
                remoteHttpError:
                  value:
                    type: https://raplp.digg.se/problems/bad-request
                    title: Invalid Request
                    status: 400
                    detail: The requested URL returned HTTP 404.
                    instance: /api/v1/validation/url
                specValidation:
                  value:
                    type: https://raplp.digg.se/problems/semantic-validation
                    title: Rule validation failed
                    status: 400
                    detail: Specifikationen innehåller strukturella eller semantiska fel
                    instance: /api/v1/validation/url
                    kind: spec-validation
                    stage: strict
                    format: yaml
                    issues:
                      - code: parser-error
                        message: Missing required property paths.
                        severity: error
                        path:
                          - paths
                    snippet: |
                      line 12: missing required property: paths
                ruleValidation:
                  value:
                    type: https://raplp.digg.se/problems/rule-validation
                    title: Rule validation failed
                    status: 400
                    detail: API-specifikationen bryter mot en eller flera regler enligt den svenska REST API-profilen.
                    instance: /api/v1/validation/url
                    kind: rule-validation
                    stage: rule-engine
                    format: yaml
                    payload:
                      result:
                        - id: AME.04
                          område: API Message
                          krav: For faltnamn i request och response body BOR camelCase eller snake_case notation anvandas.
                          allvarlighetsgrad: WARNING
                          sökväg:
                            - components
                            - schemas
                            - Example
                            - properties
                            - problemField
                          omfattning:
                            start:
                              line: 10
                              character: 8
                            end:
                              line: 10
                              character: 20
                      report:
                        - Notering: Ej godkanda regler - RAP-LP
                          regler:
                            - id: AME.04
                              område: API Message
                              status: EJ OK
        '409':
          description: URL validation is disabled by server configuration.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                featureDisabled:
                  value:
                    type: https://raplp.digg.se/problems/conflict
                    title: Conflict
                    status: 409
                    detail: This feature is currenctly dissabled due to server configuration. Contact your administrator if you think this is a misstake.
                    instance: /api/v1/validation/url
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                internalServerError:
                  value:
                    type: https://raplp.digg.se/problems/internal-server-error
                    title: Internt serverfel
                    status: 500
                    detail: Ett ovantat fel intraffade.
                    instance: /api/v1/validation/url
...
components:
  schemas:
...
    UrlValidationRequest:
      type: object
      additionalProperties: false
      properties:
        url:
          type: string
          format: uri
          description: Absolute URL pointing to a remotely hosted OpenAPI specification.
        strict:
          type: boolean
          description: Enables strict parsing and semantic validation before RAP-LP rules are executed.
          default: true
        categories:
          type: array
          description: Optional subset of rule categories to execute. If omitted, all categories are used and ForRules is added implicitly by the service.
          items:
            type: string
          examples:
            - [DokRules, UfnRules]
      required:
        - url
...
    ValidationResult:
      type: object
      additionalProperties: false
      properties:
        result:
          type: array
          items:
            $ref: '#/components/schemas/RapLpDiagnosticItem'
        report:
          type: array
          items:
            $ref: '#/components/schemas/DiagnosticReport'
      required:
        - result
        - report
...
    UrlValidationSuccessResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        stage:
          type: string
          enum:
            - rule-engine
          description: Processing stage that produced the successful response.
        format:
          $ref: '#/components/schemas/SpecFormat'
        payload:
          $ref: '#/components/schemas/ValidationResult'
      required:
        - ok
        - stage
        - format
        - payload
    SpecFormat:
      type: string
      description: Parsed OpenAPI document format.
      enum:
        - yaml
        - json
      example: yaml
    ValidationStage:
      type: string
      description: Validation stage that produced the response.
      enum:
        - strict
        - rule-engine
      example: rule-engine
    ProblemDetails:
      type: object
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
      required:
        - type
        - title
        - status
        - detail
        - instance
    ValidationProblemDetails:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          additionalProperties: true
          properties:
            kind:
              type: string
              enum:
                - spec-validation
                - rule-validation
            stage:
              $ref: '#/components/schemas/ValidationStage'
            format:
              $ref: '#/components/schemas/SpecFormat'
            issues:
              type: array
              items:
                $ref: '#/components/schemas/StrictIssue'
            snippet:
              type: string
            payload:
              $ref: '#/components/schemas/ValidationResult'
    StrictIssue:
      type: object
      additionalProperties: true
      properties:
        code:
          type: string
        message:
          type: string
        severity:
          type: string
        path:
          type: array
          items:
            type: string
      required:
        - message
    RapLpDiagnosticItem:
      type: object
      additionalProperties: true
      properties:
        id:
          type: string
          nullable: true
        område:
          type: string
          nullable: true
        krav:
          type: string
          nullable: true
        allvarlighetsgrad:
          type: string
          nullable: true
          enum:
            - HINT
            - INFORMATION
            - WARNING
            - ERROR
        sökväg:
          type: array
          items:
            type: string
          nullable: true
        omfattning:
          $ref: '#/components/schemas/Range'
        tags:
          type: array
          items:
            type: string
        relatedInformation:
          type: array
          items:
            $ref: '#/components/schemas/DiagnosticRelatedInformation'
    DiagnosticRelatedInformation:
      type: object
      additionalProperties: false
      properties:
        location:
          $ref: '#/components/schemas/Location'
        message:
          type: string
      required:
        - location
        - message
    Location:
      type: object
      additionalProperties: false
      properties:
        uri:
          type: string
          nullable: true
        range:
          $ref: '#/components/schemas/Range'
      required:
        - range
    Range:
      type: object
      additionalProperties: false
      properties:
        start:
          $ref: '#/components/schemas/Position'
        end:
          $ref: '#/components/schemas/Position'
      required:
        - start
        - end
    Position:
      type: object
      additionalProperties: false
      properties:
        line:
          type: integer
          minimum: 0
        character:
          type: integer
          minimum: 0
      required:
        - line
        - character
...
    PopulatedDiagnosticRuleInfo:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          example: UFN.05
        område:
          type: string
          example: URL Format och namngivning
        helpUrl:
          type: string
          format: uri
        krav:
          type: string
        status:
          type: string
          example: OK
      required:
        - id
        - område
        - status
    DiagnosticReport:
      type: object
      additionalProperties: false
      properties:
        Notering:
          type: string
          example: Godkanda regler - RAP-LP
        regler:
          type: array
          items:
            $ref: '#/components/schemas/PopulatedDiagnosticRuleInfo'
      required:
        - Notering
        - regler
...